🚚 Rename(#1): 쇼츠 상세 관련 dao 분리
